+++ pycaptions/captionsFormat.py
@@ -37,6 +37,81 @@ save_extensions.TTML = ".xml"
 """
 
 
+def captionsDetector(func):
+    def wrapper(content):
+        if not isinstance(content, io.IOBase):
+            if not isinstance(content, str):
+                raise ValueError("The content is not a unicode string or I/O stream.")
+            content = io.StringIO(content)
+        offset = content.tell()
+        result = func(content=content)
+        content.seek(offset)
+        return result
+    return wrapper
+
+
+def captionsWriter(extension: str, generator_type: str = None, new_line: str = "\n"):
+    def decorator(func):
+        def wrapper(self, filename: str, languages: list[str] = None, **kwargs):
+            filename = self.makeFilename(filename=filename, extension=getattr(self.extensions, extension),
+                                    languages=languages, **kwargs)
+            encoding = kwargs.get("file_encoding") or "UTF-8"
+            languages = languages or [self.default_language]
+
+            if "lines" in kwargs:
+                lines = kwargs["lines"]
+                del kwargs["lines"]
+            else:
+                lines = -1
+
+            if "new_line" in kwargs:
+                new_line = kwargs["new_line"]
+            
+            if kwargs.get("generator"):
+                generator = kwargs.get("generator")
+            else:
+                if "style" in kwargs and kwargs["style"] != "full":
+                    if kwargs["style"] == None:
+                        generator = (((new_line.join(data.get(lang=i, lines=lines, **kwargs)) for i in languages), data) for data in self)
+                    else:
+                        raise ValueError(f"Incorect argument value of style, expected 'full' or None, got {kwargs['style']}")
+                else:
+                    generator = (((getattr(data.get_style(i), generator_type)(lines=lines, **kwargs) for i in languages), data) for data in self)
+
+            try:
+                with open(filename, "w", encoding=encoding) as file:
+                    func(self=self, filename=filename, languages=languages, generator=generator, file=file, **kwargs)
+            except IOError as e:
+                print(f"I/O error({e.errno}): {e.strerror}")
+            except Exception as e:
+                print(f"Error {e}")
+            
+        return wrapper
+    return decorator
+
+def captionsReader(func):
+    """
+    Decorator for captions readers
+
+    Parameters:
+    - content (str | io.IOBase): Content of file or string
+    - languages (list[str], optional): list of languages (default self.default_language)
+    - time_offset (MicroTime, optional): Used for shifting time on read (default is 0)
+    """
+    def wrapper(self, content: str | io.IOBase, languages: list[str] = None, 
+                time_offset: MT = None, **kwargs):
+        if not isinstance(content, io.IOBase):
+            if not not isinstance(content, str):
+                raise ValueError("The content is not a unicode string or I/O stream.")
+            content = io.StringIO(content)
+        languages = languages or [self.default_language]
+        time_offset = kwargs.get("time_offset") or MT()
+        func(self, content, languages, **kwargs)
+        self.shift_time(time_offset)
+
+    return wrapper
+
+
 class CaptionsFormat:
     """
     Represents a format for handling captions in a multimedia context.
@@ -44,7 +119,7 @@ class CaptionsFormat:
     Attributes:
         extensions (FileExtensions): An instance of the FileExtensions class for managing file extensions.
 
-     Methods:
+    Methods:
         setDefaultLanguage: Set the default language for captions.
         insert: Insert a block at the specified index.
         detect: Detect the format of the captions file.
@@ -85,7 +160,7 @@ class CaptionsFormat:
         - **options: Additional keyword arguments for customization (e.g. metadata, style, ...).
         """
         self.json_version = options.get("json_version") or JSON_VERSION
-        self.time_length = MT() or time_length
+        self.time_length =  time_length or MT()
         self.file_name_or_content = file_name_or_content
         self.isFile = isFile
         self.legacyJson = legacyJson
@@ -276,13 +351,6 @@ class CaptionsFormat:
     def read(self, content: str | io.IOBase, languages: list[str] = None, **kwargs):
         raise ValueError("Not implemented")
 
-    def checkContent(self, content: str | io.IOBase, **kwargs):
-        if not isinstance(content, io.IOBase):
-            if not not isinstance(content, str):
-                raise ValueError("The content is not a unicode string or I/O stream.")
-            content = io.StringIO(content)
-        return content
-
     def save(self, filename: str, languages: list[str] = None, **kwargs):
         raise ValueError("Not implemented")
 
@@ -322,15 +390,18 @@ class CaptionsFormat:
 
     def shift_time(self, time: MT):
         for i in self:
-            i.shift_time_us(time)
+            if i.block_type == BlockType.CAPTION:
+                i.shift_time(time)
 
     def shift_start(self, time: MT):
         for i in self:
-            i.shift_start_us(time)
+            if i.block_type == BlockType.CAPTION:
+                i.shift_time(time)
 
     def shif_end(self, time: MT):
         for i in self:
-            i.shift_end_us(time)
+            if i.block_type == BlockType.CAPTION:
+                i.shift_time(time)
 
     def _loadJson(self, data, **kwargs):
         self.time_length = data["time_length"]
@@ -475,21 +546,3 @@ class CaptionsFormat:
         with open(file, "rb") as f:
             return detect_encoding(f.read()).get("encoding")
         
-    def getGenerator(self, function: str, languages: list[str], **kwargs):
-        languages = languages or [self.default_language]
-        if "lines" in kwargs:
-            lines = kwargs["lines"]
-            del kwargs["lines"]
-        else:
-            lines = -1
-        new_line = "\n"
-        if "new_line" in kwargs:
-            new_line = kwargs["new_line"]
-        
-        if "style" in kwargs and kwargs["style"] != "full":
-            if kwargs["style"] == None:
-                return (((new_line.join(data.get(lang=i, lines=lines, **kwargs)) for i in languages), data) for data in self)
-            else:
-                raise ValueError(f"Incorect argument value of style, expected 'full' or None, got {kwargs['style']}")
-        else:
-            return (((getattr(data.get_style(i), function)(lines=lines, **kwargs) for i in languages), data) for data in self)