+++ pycaptions/lrc.py
@@ -2,7 +2,7 @@ import io
 import re
 
 from .block import Block, BlockType
-from .captionsFormat import CaptionsFormat
+from .captionsFormat import CaptionsFormat, captionsDetector, captionsReader, captionsWriter
 from .microTime import MicroTime as MT
 
 
@@ -10,6 +10,7 @@ EXTENSIONS = [".lrc"]
 
 
 @staticmethod
+@captionsDetector
 def detectLRC(content: str | io.IOBase) -> bool:
     r"""
     Used to detect Synchronized Accessible Media Interchange caption format.
@@ -18,39 +19,19 @@ def detectLRC(content: str | io.IOBase) -> bool:
      - the first line starts with `[` and ends with `]` OR
      - ^\[(\d{1,3}):(\d{1,2}(?:[:.]\d{1,3})?)\]
     """
-    if not isinstance(content, io.IOBase):
-        if not isinstance(content, str):
-            raise ValueError("The content is not a unicode string or I/O stream.")
-        content = io.StringIO(content)
-
-    offset = content.tell()
     line = content.readline().strip()
     if line.startswith("[") and line.endswith("]") or re.match(r"^\[(\d{1,3}):(\d{1,2}(?:[:.]\d{1,3})?)\]", line):
-        content.seek(offset)
         return True
-    content.seek(offset)
     return False
 
-
+@captionsReader
 def readLRC(self, content: str | io.IOBase, languages: list[str] = None, **kwargs):
-    content = self.checkContent(content=content, **kwargs)
-    languages = languages or [self.default_language]
-    time_offset = kwargs.get("time_offset") or MT()
     raise ValueError("Not Implemented")
 
 
-def saveLRC(self, filename: str, languages: list[str] = None, **kwargs):
-    filename = self.makeFilename(filename=filename, extension=self.extensions.SAMI,
-                                 languages=languages, **kwargs)
-    encoding = kwargs.get("file_encoding") or "UTF-8"
-    try:
-        with open(filename, "w", encoding=encoding) as file:
-            for text, data in self.getGenerator("getLRC", languages, **kwargs):
-                pass
-    except IOError as e:
-        print(f"I/O error({e.errno}): {e.strerror}")
-    except Exception as e:
-        print(f"Error {e}")
+@captionsWriter("LRC", "getLRC")
+def saveLRC(self, filename: str, languages: list[str] = None, generator: list = None, 
+            file: io.FileIO = None, **kwargs):
     raise ValueError("Not Implemented")