+++ pycaptions/usf.py
@@ -1,7 +1,7 @@
 import io
 
 from .block import Block, BlockType
-from .captionsFormat import CaptionsFormat
+from .captionsFormat import CaptionsFormat, captionsDetector, captionsReader, captionsWriter
 from .microTime import MicroTime as MT
 
 
@@ -9,6 +9,7 @@ EXTENSIONS = [".usf"]
 
 
 @staticmethod
+@captionsDetector
 def detectUSF(content: str | io.IOBase) -> bool:
     """
     Used to detect Universal Subtitle Format caption format.
@@ -16,39 +17,17 @@ def detectUSF(content: str | io.IOBase) -> bool:
     It returns True if:
      - the first line starts with <USFSubtitles
     """
-    if not isinstance(content, io.IOBase):
-        if not isinstance(content, str):
-            raise ValueError("The content is not a unicode string or I/O stream.")
-        content = io.StringIO(content)
-
-    offset = content.tell()
     if content.readline().lstrip().startswith("<USFSubtitles"):
-        content.seek(offset)
         return True
-    content.seek(offset)
     return False
 
-
+@captionsReader
 def readUSF(self, content: str | io.IOBase, languages: list[str] = None, **kwargs):
-    content = self.checkContent(content=content, **kwargs)
-    languages = languages or [self.default_language]
-    time_offset = kwargs.get("time_offset") or MT()
     raise ValueError("Not Implemented")
 
-
-def saveUSF(self, filename: str, languages: list[str] = None, **kwargs):
-    filename = self.makeFilename(filename=filename, extension=self.extensions.SAMI,
-                                 languages=languages, **kwargs)
-    encoding = kwargs.get("file_encoding") or "UTF-8"
-    languages = languages or [self.default_language]
-    try:
-        with open(filename, "w", encoding=encoding) as file:
-            for text, data in self.getGenerator("getUSF", languages, **kwargs):
-                pass
-    except IOError as e:
-        print(f"I/O error({e.errno}): {e.strerror}")
-    except Exception as e:
-        print(f"Error {e}")
+@captionsWriter("USF", "getUSF")
+def saveUSF(self, filename: str, languages: list[str] = None, generator: list = None, 
+            file: io.FileIO = None, **kwargs):
     raise ValueError("Not Implemented")